feat: Create Seasonable module with season methods
